move lib/io/parse.ts to lib/structure/parse.ts
- Updated import paths in various files to reflect the new structure
- move `detect_structure_type` function from lib/io/index.ts to lib/structure/parse.ts
