+++ src/lib/structure/parse.ts
@@ -4,7 +4,7 @@ import type { Matrix3x3 } from '$lib/math'
 import * as math from '$lib/math'
 import type { AnyStructure, PymatgenStructure } from '$lib/structure'
 import { load as yaml_load } from 'js-yaml'
-import { COMPRESSION_EXTENSIONS } from './decompress'
+import { COMPRESSION_EXTENSIONS } from '../io/decompress.ts'
 
 export interface ParsedStructure {
   sites: Site[]
@@ -1346,4 +1346,47 @@ export function is_structure_file(filename: string): boolean {
   }
 
   return false
+}
+
+export const detect_structure_type = (
+  filename: string,
+  content: string,
+): `crystal` | `molecule` | `unknown` => {
+  const lower_filename = filename.toLowerCase()
+
+  if (filename.endsWith(`.json`)) {
+    try {
+      const parsed = JSON.parse(content)
+      // Check for OPTIMADE JSON format (has data.attributes.lattice_vectors)
+      if (parsed.data?.attributes?.lattice_vectors) return `crystal`
+      // Check for dimension_types (OPTIMADE format)
+      if (parsed.data?.attributes?.dimension_types?.some((dim: number) => dim > 0)) {
+        return `crystal`
+      }
+      // Check for pymatgen JSON format (has lattice property)
+      if (parsed.lattice) return `crystal`
+      // Check for other crystal indicators
+      if (parsed.data?.attributes?.nperiodic_dimensions > 0) return `crystal`
+
+      return `molecule`
+    } catch {
+      return `unknown`
+    }
+  }
+
+  if (lower_filename.endsWith(`.cif`)) return `crystal`
+  if (lower_filename.includes(`poscar`) || filename === `POSCAR`) return `crystal`
+
+  if (lower_filename.endsWith(`.yaml`) || lower_filename.endsWith(`.yml`)) {
+    return content.includes(`phono3py:`) || content.includes(`phonopy:`)
+      ? `crystal`
+      : `unknown`
+  }
+
+  if (lower_filename.match(/\.(xyz|extxyz)(?:\.(?:gz|gzip|zip|bz2|xz))?$/)) {
+    const lines = content.trim().split(/\r?\n/)
+    return lines.length >= 2 && lines[1].includes(`Lattice=`) ? `crystal` : `molecule`
+  }
+
+  return `unknown`
 }
