+++ src/lib/trajectory/parse.ts
@@ -3,9 +3,9 @@ import type { AnyStructure, ElementSymbol, Vec3 } from '$lib'
 import { is_binary } from '$lib'
 import { atomic_number_to_symbol } from '$lib/composition/parse'
 import { COMPRESSION_EXTENSIONS } from '$lib/io/decompress'
-import { parse_xyz } from '$lib/io/parse'
 import type { Matrix3x3 } from '$lib/math'
 import * as math from '$lib/math'
+import { parse_xyz } from '$lib/structure/parse'
 import type { Dataset, Entity, Group } from 'h5wasm'
 import * as h5wasm from 'h5wasm'
 import type {
