+++ src/site/structures/index.ts
@@ -1,5 +1,9 @@
 import type { FileInfo, PymatgenStructure } from '$lib'
-import { detect_structure_type, is_optimade_json, parse_optimade_json } from '$lib/io'
+import {
+  detect_structure_type,
+  is_optimade_json,
+  parse_optimade_json,
+} from '$lib/structure/parse'
 
 export const structures = Object.entries( // JSON structure files (OPTIMADE/pymatgen format) as JS objects
   import.meta.glob(`./*.json`, {
