+++ tests/vitest/trajectory/streaming.test.ts
@@ -427,7 +427,7 @@ describe(`Trajectory Streaming`, () => {
       const max_time = Math.max(early_access, middle_access, late_access)
       const min_time = Math.min(early_access, middle_access, late_access)
 
-      expect(max_time / min_time).toBeLessThan(3) // Should not scale linearly (was toBeLessThan(2) but too flaky)
+      expect(max_time / min_time).toBeLessThan(4) // Should not scale linearly (was toBeLessThan(2) but too flaky)
     })
 
     it(`should extract metadata faster than loading full frames`, async () => {
